Add color code by default to title
Although there is a color that can be used for the title, it is not used
and is set as null in MsgConfig.sendPopup. While I could've made it use
the value, there is no config option to set the color of the title
message. And honestly, I see no need to do that for a single message.